Optimize logic

+++ DarkMoonProject/Assets/Game/Scripts/Data/AbilityData.cs
@@ -8,7 +8,7 @@ public class AbilityData {
 
     public int id = 1;
 
-    public double baseValue = 0;
+    public int baseValue = 0;
 
     public string abilityName = "";
